Load prerequisite data in separate component
