+++ mathesar_ui/src/api/types/dataFiles.ts
@@ -1,0 +1,4 @@
+export interface DataFile {
+  id: number;
+  header: boolean;
+}
